requirements.txt
flask>=3.0
spotipy>=2.23
openai>=1.0
python-dotenv>=1.0
google-genai>=1.0
